requirements.txt
pwntools
colorama
